Aiogram3New/requirements.txt
asyncpg
aiogram
APScheduler
python-decouple


